re-organize gamechangerml/src/search/doc_compare: - move DocCompareSentenceSearcher from model.py to new sentence_searcher.py - move DocCompareSentenceEncoder from model.py to new sentence_encoder.py - move SimilarityRanker from model.py to new similarity_ranker.py - add DocCompareSentenceSearcher, DocCompareSentenceEncoder, and SimilarityRanker to the __init__.py - update references to these classes
